pyproject.toml
[tool.poetry]
name = "access-exodus"
version = "0.1.0"
description = ""
authors = ["Matthew <mchen146@ucsc.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
pyodbc = "^5.0.1"
aioodbc = "^0.5.0"
pyparsing = "^3.1.1"
pyedifice = "^0.2.1"
typing-extensions = "^4.8.0"
pyside6 = "^6.6.0"
typeguard = "^4.1.5"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
